Add checking to the chess game
It is done in such a way that:
The king cannot walk into an attack of  an enemy piece.
If king is being attacked the king either has limited
move options or has none.
If the king is being attacked, the player can only move
pieces that block the enemy attack on the king.
If enemy pieces action moves are in the way of castling,
the player should not be able to castle in the direction
which is being attacked.
When the player is checked, it should be highlighted so
it is easier to see.