Add promoted checkers piece capture moves
Added the available moves to the promoted checkers
piece.

Made it so that the promoted checkers piece
can take an enemy piece from anywehere on the same
diagonal line. While taking an enemy piece the
promoted piece can land anywhere on the diagonal
line behind the captured enemy piece as long as
there are no obstructions.